% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temp.R
\name{plot_temp}
\alias{plot_temp}
\title{Faceted plot of water temperature}
\usage{
plot_temp(data, depths = 0.5, tzone = "America/New_York", ylims = c(-5, 35))
}
\arguments{
\item{data}{data frame to be plotted}

\item{depths}{vector of depths to be facets}

\item{tzone}{time zone for datetime in plot}

\item{ylims}{vector of c(lower,upper) bounds of temperature on plot}
}
\value{

}
\description{
Faceted plot of water temperature
}
