% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temp_single_panel.R
\name{plot_temp_single_panel}
\alias{plot_temp_single_panel}
\title{Temperature plot with all depths on a single panel}
\usage{
plot_temp_single_panel(
  data,
  depths = 0.5,
  tzone = "America/New_York",
  ylims = c(-5, 35),
  site_name = ""
)
}
\arguments{
\item{data}{data frame of scores}

\item{depths}{vector of depths to be facets}

\item{tzone}{time zone for datetime in plot}

\item{ylims}{vector of c(lower,upper) bounds of temperature on plot}

\item{site_name}{name of site}
}
\value{

}
\description{
Temperature plot with all depths on a single panel
}
